+++ railties/test/application/configuration_test.rb
@@ -965,5 +965,29 @@ module ApplicationTests
 
       assert db_config.is_a?(Hash)
     end
+
+    test 'config.mail_preview_enabled defaults to true in development' do
+      Rails.env = "development"
+      require "#{app_path}/config/environment"
+
+      assert Rails.application.config.action_mailer.preview_enabled
+    end
+
+    test 'config.mail_preview_enabled defaults to false in production' do
+      Rails.env = "production"
+      require "#{app_path}/config/environment"
+
+      assert_equal Rails.application.config.action_mailer.preview_enabled, false
+    end
+
+    test 'config.mail_preview_enabled can be set in the configuration file' do
+      Rails.env = "production"
+      add_to_config <<-RUBY
+        config.action_mailer.preview_enabled = true
+      RUBY
+      require "#{app_path}/config/environment"
+
+      assert_equal Rails.application.config.action_mailer.preview_enabled, true
+    end
   end
 end